SignalFunctions

Description

Signal functions are stateful, reactive, and time-dependent entities which respond to inputs by producing outputs. Signal functions are typed using signal vectors, exported by the SignalVectors module. A signal vector is formed from some combination of signals, events, and empty indices. Signal inputs update only at time steps, while events may have occurrences at any time, including between time steps. Event occurrences form a preorder with respect to indicated occurrence times, (the reflexivity condition is not satisfied).

To program with signal functions, primitive signal functions may be combined using the routing combinators to connect inputs and outputs, as well as the reactivity combinator switch to permit signal functions to change behavior, and the feedback combinator loop to permit signal functions to respond to their own output.

In this documentation the term given signal function(s) refers to signal functions passed as parameters to signal function combinators, while (in that context) signal function will refer to the signal function produced by a combinator.

Datatypes

data SF

Signal functions: The first type parameter distinguishes between uninitialized (just specificied) and initialized (ready to respond to input and time) signal functions. The second and third type parameters are the input and output signal vectors, respectively.

data NonInitialized

Signal function not yet running

data Initialized

Signal function running or suspended

Basic signal functions

identity :: SF NonInitialized sv sv

Identity signal function: reproduce the input exactly as the output.

Lifting

pureSignalTransformer :: (a -> b) -> SF NonInitialized (SVSignal a) (SVSignal b)

Apply the given function to every sample of a signal

pureEventTransformer :: (a -> b) -> SF NonInitialized (SVEvent a) (SVEvent b)

Apply the given function to each occurrence of an event

Composition and Routing

(>>>) :: SF NonInitialized svIn svBetween -> SF NonInitialized svBetween svOut -> SF NonInitialized svIn svOut

Produce a new signal function where the output of the first signal function is used as the input to the second signal function

first :: SF NonInitialized svIn svOut -> SF NonInitialized (SVAppend svIn sv) (SVAppend svOut sv)

Produce a new signal function where the left side of the input is used as input for the given signal function, and the right side of the input is combined unchanged with the output of the given signal function

second :: SF NonInitialized svIn svOut -> SF NonInitialized (SVAppend sv svIn) (SVAppend sv svOut)

Similar to first, except the right side of the signal vector is used as input for the given signal function, and the left side is combined unchanged with the output.

swap :: SF NonInitialized (SVAppend sv1 sv2) (SVAppend sv2 sv1)

Produce the left side of the input as the right side of the output, and vice versâ.

copy :: SF NonInitialized sv (SVAppend sv sv)

Produce the input as left and right sides of the output

ignore :: SF NonInitialized sv SVEmpty

Discard the input entirely and produce the left side of the output

cancelLeft :: SF NonInitialized (SVAppend SVEmpty sv) sv

Discard the empty left side of the input and use the right side as output.

uncancelLeft :: SF NonInitialized sv (SVAppend SVEmpty sv)

Use the input as the right side of the output and leave the left side empty

cancelRight :: SF NonInitialized (SVAppend sv SVEmpty) sv

Discard the empty right side of the input and use the left side as the output.

uncancelRight :: SF NonInitialized sv (SVAppend sv SVEmpty)

Use the input as the left side of the output and leave the right side empty

associate :: SF NonInitialized (SVAppend (SVAppend sv1 sv2) sv3) (SVAppend sv1 (SVAppend sv2 sv3))

Take an input where the left two subvectors are associated and produce an output with the same subvectors, but where the right two subvectors are associated.

unassociate :: SF NonInitialized (SVAppend sv1 (SVAppend sv2 sv3)) (SVAppend (SVAppend sv1 sv2) sv3)

Take an input where the right two subvectors are associated and produce an output with the same subvectors, but where the left two subvectors are associated.

Reactivity

switch :: SF NonInitialized svIn (SVAppend svOut (SVEvent (SF NonInitialized svIn svOut))) -> SF NonInitialized svIn svOut

Provide the input as input to the given signal function, producing the left side of the given signal function's output as output. Upon the first occurrence of the event on the right side of the given signal function's output, act as the signal function contained by the occurrence. Note that the if a switch event occurs in between time samples, the switch function will combine the signal deltas produced by the initialization of the new signal function and by the first time step of the new signal function.

Feedback

loop :: SF NonInitialized (SVAppend svIn svLoop) (SVAppend svOut svLoop) -> SF NonInitialized svIn svOut

Provide the input as the left input to the given signal function. Produce the left output of the given signal function as the output. Use the right output of the given signal function as the right input to the given signal function, thus providing a means for a signal function to act on its own output and encode a feedback loop. This function is decoupled for signals, meaning that a signal function will produce output that depends on the signal input at the previous time step, but not for events, meaning that feedback events will have an immediate effect. It is the programmer's responsibility to ensure that an feedback event does not result in an infinite set of resulting events.

Time dependence

time :: SF NonInitialized SVEmpty (SVSignal Double)

With an empty input, produce the time since the signal function began running as a signal output.

delay :: Double -> SF NonInitialized (SVAppend (SVEvent a) (SVEvent Double)) (SVEvent a)

Maintain a time delay for events. The initial argument is the initial time delay. Left input event occurrences are stored as state together with the time delay at occurrence time. After that delay has expired, a corresponding event will be produced on the output. Right input event occurrences update the delay time. Note that this means that later input events may be output before prior input events if the delay time is shortened in the intervening time.

class TimeIntegrate i

Class of values integrateable with respect to time

integrate :: TimeIntegrate i => SF NonInitialized (SVSignal i) (SVSignal i)

Produce as output the rectangle rule integration of the input with respect to time.

Joining

union :: SF NonInitialized (SVAppend (SVEvent a) (SVEvent a)) (SVEvent a)

Produce an event occurrence on the output corresponding to an event occurrence on either input

combineSignals :: ((a, b) -> c) -> SF NonInitialized (SVAppend (SVSignal a) (SVSignal b)) (SVSignal c)

Combine both inputs at each point in time using the given function.

Evaluation

data SignalDelta sv

Descriptors of signal changes for a time step

sd :: a -> SignalDelta (SVSignal a)

Create a signal delta for a singleton input

sdLeft :: SignalDelta svl -> SignalDelta (SVAppend svl svr)

Use this signal delta for only the left input

sdRight :: SignalDelta svr -> SignalDelta (SVAppend svl svr)

Use this signal delta for only the right input

sdBoth :: SignalDelta svl -> SignalDelta svr -> SignalDelta (SVAppend svl svr)

Use these signal deltas for the left and right inputs, respectively

data EventOccurrence sv

Descriptors of event occurrences

eo :: a -> EventOccurrence (SVEvent a)

Create an event occurrence to push

eoLeft :: EventOccurrence svl -> EventOccurrence (SVAppend svl svr)

Route an event occurrence left

eoRight :: EventOccurrence svr -> EventOccurrence (SVAppend svl svr)

Route an event occurrence right

data HandlerSet sv m

A set of actions to handle SF outputs. These outputs may be the components of a signal delta or may be event occurrences.

sdHS :: (a -> m ()) -> HandlerSet (SVSignal a) m

A handler for a signal delta output

eoHS :: (a -> m ()) -> HandlerSet (SVEvent a) m

A handler for an event occurrence output

hsLeft :: HandlerSet svl m -> HandlerSet (SVAppend svl svr) m

Use the given handler for the left side of the output, and do not handle the right.

hsRight :: HandlerSet svr m -> HandlerSet (SVAppend svl svr) m

Use the given handler for the right side of the output, and do not handle left side.

hsBoth :: HandlerSet svl m -> HandlerSet svr m -> HandlerSet (SVAppend svl svr) m

Use the given handlers for the left and right sides of the output, respectively.

data SFEvalT svIn svOut m a

A monad for evaluating signal functions

type SFEvalIO svIn svOut a = SFEvalT svIn svOut IO a

The most likely use case:

data SFEvalState m svIn svOut

SFEvalState

runSFEvalT :: SFEvalT svIn svOut m a -> SFEvalState m svIn svOut -> m (a, SFEvalState m svIn svOut)

Evaluate a signal function, whose current state is stored in an SFEvalState.

initSFEval :: HandlerSet svOut m -> SF NonInitialized svIn svOut -> SFEvalState m svIn svOut

Initialize the evaluation of a signal function, producing an SFEvalState.

push :: Monad m => EventOccurrence svIn -> SFEvalT svIn svOut m ()

Push an event occurrence into the input of the signal function

sample :: Monad m => Double -> SignalDelta svIn -> SFEvalT svIn svOut m ()

Sample the signal function