SignalVectors

Description

This module defines signal vectors, a type-level representation of the input and output structure of signal functions, and value representations typed with signal vectors for use in implementing signal functions.

Type Representation of Signal Vectors

A signal vector is a type which describes the structure of the input or output of a signal function. Signal vectors do not have value members, but by representing signal functions and other implementation types using GADTs, we can include signal vectors in the types of signal functions and their implementations.

data SVEmpty

Empty signal vector phantom type

data SVSignal

Singleton signal signal vector phantom type

data SVEvent

Singleton event signal vector phantom type

data SVAppend

Combine two signal vector phantom types

Signal Vector Element Functors

These newtypes are used as higher-kinded type parameters to the constructors of signal vector indices and signal vector memories, which both apply the functor to each element of the signal vector

newtype Id a

"Identity" type constructor, for use as a functor on the element types of a signal vector

Constructors

Id a

newtype To b a

Newtype representing a function "to" a value of the first type parameter, from a value of the second parameter. Essentially flip for the arrow type.

Constructors

To (a -> b)

Value Representations of Signal Vectors

These types are used in the implementation of signal functions to represent values for inputs and outputs. The first parameter to each type is higher-kinded, and is applied to the type at an index of a signal vector to produce the actual type contained in the representation. This permits, for instance, the re-use of these representations for storing output handlers in the monadic evaluator.

Signal Vector Indices

Signal vector indices are representations of individual values at a particular index of a signal vector. They take the type of the whole signal vector, while dynamically representing which index the value corresponds to, and statically ensuring that this value matches the type. This representation is used internally to the implementation of signal functions (module SignalFunctions) and is not manipulated by users.

Datatype

data SVIndex where

A single value at an index of a signal vector

Constructors

SVISignal :: p a -> SVIndex p (SVSignal a)

SVIEvent :: p a -> SVIndex p (SVEvent a)

SVILeft :: SVIndex p svl -> SVIndex p (SVAppend svl svr)

SVIRight :: SVIndex p svr -> SVIndex p (SVAppend svl svr)

Helper Functions

indexIsEvent :: SVIndex p sv -> Bool

Is this particular index to an event?

indexIsSignal :: SVIndex p sv -> Bool

Is this particular index to a signal?

splitIndices :: [SVIndex p (SVAppend svl svr)] -> ([SVIndex p svl], [SVIndex p svr])

Split a list of indices into lists of left and right indices

Signal Memories

Signal memories store values for a subset of all indices in a signal vector. Signal memories are used in the implementation of signal functions (module SignalFunctions) but may also be manipulated by the user for the specification of properties corresponding to evaluated signal functions. For instance, the monadic evaluation interface uses signal memories to specify output handlers for the signal function being evaluated.

Datatypes

data SMemory where

A partial collection of values at indices of a signal vector

Constructors

SMEmpty :: SMemory p sv

SMSignal :: p a -> SMemory p (SVSignal a)

SMEvent :: p a -> SMemory p (SVEvent a)

SMBoth :: SMemory p svl -> SMemory p svr -> SMemory p (SVAppend svl svr)

Helper Functions

emptySMemory :: SMemory p sv

An empty signal memory

updateSMemory :: SMemory p sv -> SVIndex p sv -> SMemory p sv

Update signal memory with a signal vector index

updateWithSMemory :: SMemory p v -> SMemory p v -> SMemory p v

Merge two signal memories, given priority

combineSignalMemory :: SMemory p vl -> SMemory p vr -> SMemory p (SVAppend vl vr)

Combine two signal memories

toIndices :: SMemory p v -> [SVIndex p v]

Create a list of signal vector indices from a signal memory

applySMTo :: SMemory (To a) sv -> [SVIndex Id sv] -> [a]

Apply a signal memory holding functions to signal indices holding values

applySMToSM :: SMemory (To a) sv -> SMemory Id sv -> [a]

Apply a signal memory holding functions to a signal memory holding values, producing a list of values.